Implement RTCRtpTransceiverManager for improved transceiver management and codec negotiation
